requirements.txt
python-dotenv==1.0.0
openai==1.12.0
faiss-cpu==1.7.4
numpy==1.24.3
python-docx==1.0.1
PyPDF2==3.0.1
sentence-transformers==2.3.1